+++ src/descriptors/keys.rs
@@ -140,33 +140,6 @@ impl ToPublicKey for DerivedPublicKey {
     }
 }
 
-/// We require the descriptor key to:
-///  - Be deriveable (to contain a wildcard)
-///  - Be multipath (to contain a step in the derivation path with multiple indexes)
-///  - The multipath step to only contain two indexes, 0 and 1.
-///  - Be 'signable' by an external signer (to contain an origin)
-pub fn is_valid_desc_key(key: &descriptor::DescriptorPublicKey) -> bool {
-    match *key {
-        descriptor::DescriptorPublicKey::Single(..) | descriptor::DescriptorPublicKey::XPub(..) => {
-            false
-        }
-        descriptor::DescriptorPublicKey::MultiXPub(ref xpub) => {
-            let der_paths = xpub.derivation_paths.paths();
-            // Rust-miniscript enforces BIP389 which states that all paths must have the same len.
-            let len = der_paths.get(0).expect("Cannot be empty").len();
-            // Technically the xpub could be for the master xpub and not have an origin. But it's
-            // no unlikely (and easily fixable) while users shooting themselves in the foot by
-            // forgetting to provide the origin is so likely that it's worth ruling out xpubs
-            // without origin entirely.
-            xpub.origin.is_some()
-                && xpub.wildcard == descriptor::Wildcard::Unhardened
-                && der_paths.len() == 2
-                && der_paths[0][len - 1] == 0.into()
-                && der_paths[1][len - 1] == 1.into()
-        }
-    }
-}
-
 /// The keys in one of the two spending paths of a Liana descriptor.
 /// May either be a single key, or between 2 and 20 keys along with a threshold (between two and
 /// the number of keys).
